Rewrite Question bank search to include competency

+++ models/questions.py
@@ -5,7 +5,7 @@ db.define_table(
     Field("chapter", type="string"),  # matches chapter_label, not name
     Field("subchapter", type="string"),  # matches sub_chapter_label, not name
     Field("author", type="string"),
-    Field("difficulty", type="integer"),
+    Field("difficulty", type="float"),
     Field("question", type="text"),
     Field("timestamp", type="datetime"),
     Field("question_type", type="string"),
